disallow embedded executor for remote deployments
