+++ packages/core/src/actions/index.ts
@@ -1,3 +1,4 @@
 export * from './bundle'
 export * from './types'
 export * from './artifacts'
+export * from './execute'
