+++ packages/plugins/chugsplash/SimpleStorage.config.ts
@@ -3,7 +3,7 @@ import { ChugSplashConfig } from '@chugsplash/core'
 const config: ChugSplashConfig = {
   // Configuration options for the project:
   options: {
-    projectName: 'My First Project',
+    projectName: 'executor refactor 3',
     projectOwner: '0xf39fd6e51aad88f6f4ce6ab8827279cfffb92266',
   },
   // Below, we define all of the contracts in the deployment along with their state variables.
